Event image field fixed
