+++ mocacares_backend/event_platform/encoder.py
@@ -45,7 +45,7 @@ class EventSummaryEncoder(DjangoJSONEncoder):
                 'time_type': '1',
                 "content": "hellodhjsds",
                 "title": obj.title,
-                "img": obj.img[1:],
+                "img": obj.img.image_url[1:],
                 "desrc": obj.description,
                 "add": obj.address,
                 "type": obj.event_type.pk,
@@ -66,7 +66,7 @@ class EventDetailEncoder(DjangoJSONEncoder):
                 'hour_end': datetime_to_time_str(obj.end_time),
                 'time_type': '1',
                 "title": obj.title,
-                "img": obj.img[1:],
+                "img": obj.img.image_url[1:],
                 "desrc": obj.description,
                 "add": obj.address,
                 "type": obj.event_type.pk,
