+++ mocacares_backend/event_platform/models/models.py
@@ -1,8 +1,25 @@
 import os
 import shutil
 
+from django.dispatch import receiver
 from django.db import models
 from .models_user import User
+
+
+class UploadedImage(models.Model):
+    image = models.ImageField()
+
+    @property
+    def image_url(self):
+        return self.image.url
+
+
+@receiver(models.signals.pre_delete, sender=UploadedImage)
+def delete_local_file(sender, instance, **kwargs):
+    """
+    delete the corresponding image file
+    """
+    instance.image.delete()
 
 
 class EventType(models.Model):
@@ -13,7 +30,7 @@ class Event(models.Model):
     title = models.CharField(max_length=1028)
     description = models.TextField()
     address = models.TextField()
-    img = models.CharField(max_length=256)
+    img = models.OneToOneField(UploadedImage, on_delete=models.CASCADE)
     start_time = models.DateTimeField()
     end_time = models.DateTimeField()
     event_type = models.ForeignKey(EventType, on_delete=models.CASCADE)
